Fix command refresh
